2022/11/29 13:21

+++ 0x10_Dynamic Programming[DP]/꺾이지 않는 마음_26142.cpp
@@ -15,7 +15,7 @@ int main (void)
 	for(int i = 0; i < n; i++)
 		cin >> day[i] >> dr_len[i];
 
-	sort(dr_len, dr_len + n, greater<int>());
+	sort(dr_len, dr_len + n, greater<long long>());
 	res[0] = dr_len[0];
 	for(int i = 1; i < n; i++)
 	{
